[refactor] 테스트 파일을 패키지 별로 구분
