+++ src/test/java/lotto/controller/LottoManagerTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.controller;
 
 import lotto.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
@@ -8,7 +8,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class LottoTest {
+class LottoManagerTest {
     @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
     @Test
     void createLottoByOverSize() {
